Сделал расширение для UIViewControllera
