+++ UIKitTemplate/UIKitTemplate/SceneDelegat.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
         window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = ViewController()
+        window?.rootViewController = StartViewController()
         window?.makeKeyAndVisible()
     }
 }
